[v1.0.0] [fix: TimerUsecase SRP 위반, 리팩토링 ]
